Moved the risk function parsers into commonlib.nrml
